requirements.txt
azure-core==1.26.2
azure-functions==1.12.0
azure-storage-blob==12.14.1
azure-storage-queue==12.5.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
idna==3.4
isodate==0.6.1
msrest==0.7.1
numpy==1.24.1
oauthlib==3.2.2
pandas==1.5.3
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.14
